Organized project directory, created file to export H2 to CSV, create main file to handle scripts, Included Maven Files
Created ExportData.java to export our H2 table data to csv files. Files will be imported into sqlite.

Realized the other commits didn't have the Maven files included so I added those to this commit.

Organized project directory for better structure and ensure the database was created in the correct place.

Had to manually download H2 db as it was struggling to automatically download when the project is being built.

Instructions on how to build the project, manually download the H2 database, and run the application will be included in the readme once I'm done implementing everything. To test from scratch, delete the prehab/java_backend_migration/target/ folder.